app/backend/requirements.txt
fastapi
uvicorn
weaviate-client[agents]
python-dotenv
